+++ packages/adapter-shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chat-adapter/shared
 
+## 4.13.4
+
+### Patch Changes
+
+- Updated dependencies [716ce2a]
+  - chat@4.13.4
+
 ## 4.13.3
 
 ### Patch Changes